fix(gnoland): speed test & rework directory
