+++ gno.land/pkg/integration/testdata/adduserfrom.txtar
@@ -27,7 +27,7 @@ stdout '  "BaseAccount": {'
 stdout '    "address": "g1mtmrdmqfu0aryqfl4aw65n35haw2wdjkh5p4cp",'
 stdout '    "coins": "10000000ugnot",'
 stdout '    "public_key": null,'
-stdout '    "account_number": "59",'
+stdout '    "account_number": "58",'
 stdout '    "sequence": "0"'
 stdout '  }'
 stdout '}'
