+++ gno.land/pkg/integration/testdata/gnoland.txtar
@@ -28,16 +28,16 @@ cmp stderr gnoland-already-stop.stderr.golden
 
 -- gnoland-no-arguments.stdout.golden --
 -- gnoland-no-arguments.stderr.golden --
-"gnoland" error: syntax: gnoland [start|stop|restart]
+"gnoland" error: exit status 1
 -- gnoland-start.stdout.golden --
 node started successfully
 -- gnoland-start.stderr.golden --
 -- gnoland-already-start.stdout.golden --
+node started successfully
 -- gnoland-already-start.stderr.golden --
-"gnoland start" error: node already started
 -- gnoland-stop.stdout.golden --
 node stopped successfully
 -- gnoland-stop.stderr.golden --
 -- gnoland-already-stop.stdout.golden --
+node stopped successfully
 -- gnoland-already-stop.stderr.golden --
-"gnoland stop" error: node not started cannot be stopped
